:recycle: Move to top level reference for MockReq
